细节优化

+++ static/data/delete.php
@@ -13,6 +13,13 @@ $input = json_decode($input);
 
 settype($input, 'array');
 extract($input);
+$res = [];
+if($topic_id == null || $user_id == null){
+    $res['status'] = 0;
+    $res['info'] = '删除失败';
+    $res['data'] = '';
+    exit(json_encode($res));
+}
 $sql = "DELETE FROM t_topic WHERE topic_id=$topic_id AND user_id=$user_id";
 $result = mysqli_query($conn,$sql);
 if($result === true ){
